LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-transition-attr-expected.txt

Anchor1
Anchor2

FAIL Transition when the anchor attribute changes assert_equals: expected 0 but got 140
